create new file kernel_interface.py
